Restructure production checks
