+++ front/temporal/production_checks/activities.ts
@@ -1,14 +1,14 @@
 import { Context } from "@temporalio/activity";
 import { v4 as uuidv4 } from "uuid";
 
+import { checkActiveWorkflows } from "@app/lib/production_checks/checks/check_active_workflows_for_connectors";
+import { checkNotionActiveWorkflows } from "@app/lib/production_checks/checks/check_notion_active_workflows";
+import { managedDataSourceGCGdriveCheck } from "@app/lib/production_checks/checks/managed_data_source_gdrive_gc";
+import { managedDataSourceGdriveWebhooksCheck } from "@app/lib/production_checks/checks/managed_data_source_gdrive_webhooks";
+import { nangoConnectionIdCleanupSlack } from "@app/lib/production_checks/checks/nango_connection_id_cleanup_slack";
+import { scrubDeletedCoreDocumentVersionsCheck } from "@app/lib/production_checks/checks/scrub_deleted_core_document_versions";
+import type { Check } from "@app/lib/production_checks/types";
 import mainLogger from "@app/logger/logger";
-import { checkActiveWorkflows } from "@app/temporal/production_checks/checks/check_active_workflows_for_connectors";
-import { checkNotionActiveWorkflows } from "@app/temporal/production_checks/checks/check_notion_active_workflows";
-import { managedDataSourceGCGdriveCheck } from "@app/temporal/production_checks/checks/managed_data_source_gdrive_gc";
-import { managedDataSourceGdriveWebhooksCheck } from "@app/temporal/production_checks/checks/managed_data_source_gdrive_webhooks";
-import { nangoConnectionIdCleanupSlack } from "@app/temporal/production_checks/checks/nango_connection_id_cleanup_slack";
-import { scrubDeletedCoreDocumentVersionsCheck } from "@app/temporal/production_checks/checks/scrub_deleted_core_document_versions";
-import type { Check } from "@app/temporal/production_checks/types/check";
 
 export async function runAllChecksActivity() {
   const checks: Check[] = [
